feat: add typescript declaration file
This commit adds the typescript declaration file. It further upgrades
parcel to version 2. This needed to be done to get rid of the `Invalid
Version: undefined` error (see
https://github.com/parcel-bundler/parcel/issues/5943). While doing so I
also had to change the react import in the `example/index.js` due to a parcel
bug (see https://github.com/parcel-bundler/parcel/issues/1199).